+++ docs/CHANGES.rst
@@ -5,6 +5,8 @@ Changes
 2.5 (unreleased)
 ----------------
 
+- Add patrimony tabs to preliminary notice [URB-2617]
+  [jchandelle]
 
 - Add parameter to autocomplete to search with exact match [URB-2696]
   [jchandelle]